requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
sqlalchemy>=2.0.35,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
pgvector>=0.3.6,<1.0.0
pydantic-settings>=2.6.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.9,<1.0.0
email-validator>=2.2.0,<3.0.0
groq>=0.30.0,<1.0.0
google-genai>=1.20.0,<2.0.0
langfuse>=3.0.0,<4.0.0
httpx>=0.28.0,<1.0.0
pytest>=8.3.0,<9.0.0
